four correct answers

+++ Exercise/Exercise2/my_image.h
@@ -5,8 +5,8 @@ typedef unsigned char Channel; // 把unsigned char这个数据类型换个名字
 // unsigned char 只有正数, 表示无符号字符类型 [0,255]
 // signed char, 表示有符号字符类型 [-128,127]
 
-const int MAX_WIDTH = 64;
-const int MAX_HEIGHT = 64;
+const size_t MAX_WIDTH = 64;
+const size_t MAX_HEIGHT = 64;
 struct Vec2
 {
     size_t _x, _y; // to represent a 2D vector, 无符号整数类型变量, 用来表示对象的大小, 以字节数为单位
